requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
python-dotenv==1.0.1
pyyaml==6.0.2
SQLAlchemy>=2.0
psycopg2-binary>=2.9
